Fix 12 GCC/Linux compatibility issues and verify cross-platform build
Linux build validated on GCC 13.3.0 (Ubuntu 24.04 WSL2): 822/822 targets
compile, 735/742 tests pass (7 GCC coroutine symmetric transfer bugs),
4/6 examples work E2E against Docker Temporal.

Key fixes:
- Force FetchContent protobuf to avoid system protobuf namespace collision
- Make proto-generated and protobuf library includes SYSTEM to suppress warnings
- Add missing #include <memory> in task_completion_source.h
- Move Workflow::decode_value template after WorkflowContext definition (GCC
  checks templates at definition time, not instantiation)
- Fix char-to-unsigned-char sign conversions in base64 encode/decode
- Fix 7 int-to-size_t sign conversions in protobuf _size() to reserve() calls
- Add -Wno-missing-field-initializers for C++20 designated initializers
- Add [[maybe_unused]] to unused test helper functions

+++ cpp/include/temporalio/converters/payload_conversion.h
@@ -57,7 +57,7 @@ inline temporal::api::common::v1::Payloads to_proto_payloads(
 inline std::vector<Payload> from_proto_payloads(
     const temporal::api::common::v1::Payloads& proto) {
     std::vector<Payload> payloads;
-    payloads.reserve(proto.payloads_size());
+    payloads.reserve(static_cast<size_t>(proto.payloads_size()));
     for (const auto& p : proto.payloads()) {
         payloads.push_back(from_proto_payload(p));
     }
